Login/Register CSS
CSS & removed Header/Nav/Footer from layout to match design
