+++ src/pages/ActivityDashboardPage.jsx
@@ -2,11 +2,18 @@ import React, { useState } from "react";
 import EventForm from "../components/EventForm";
 import EventList from "../components/EventList";
 
+import Header from '../components/Header/Header.jsx';
+import Footer from '../components/Footer/Footer.jsx';
+import Navigation from '../components/Navigation/Navigation.jsx';
+
 const ActivityDashboardPage = () => {
     const [showForm, setShowForm] = useState(false);
     const [refresh, setRefresh] = useState(false);
 
     return (
+        <div>
+        <Header />
+        <Navigation />
         <div className="dashboard-container">
             <h1>Activity Dashboard</h1>
             <p>Manage and create activity posts.</p>
@@ -24,6 +31,8 @@ const ActivityDashboardPage = () => {
             {/* List of Events */}
             <h2>Recent Activities</h2>
             <EventList key={refresh} />
+        </div>
+        <Footer />
         </div>
     );
 };
