+++ src/pages/ResourcesPage.jsx
@@ -1,7 +1,14 @@
 import React from 'react';
 
+import Header from '../components/Header/Header.jsx';
+import Footer from '../components/Footer/Footer.jsx';
+import Navigation from '../components/Navigation/Navigation.jsx';
+
 const ResourcesPage = () => {
   return (
+    <div>
+    <Header />
+    <Navigation />
     <div className="resources-container">
       <h1>Resources & Guidance</h1>
       <p>Access helpful resources for healthcare training.</p>
@@ -10,6 +17,8 @@ const ResourcesPage = () => {
         <li>🎥 Video Tutorials for Medical Staff</li>
         <li>🔗 NHS Best Practices & Policies</li>
       </ul>
+    </div>
+    <Footer />
     </div>
   );
 };
